JoinPushTransitivePredicatesRule ready to ship
Added knobs for three added rules and corresponding tests in
QueryCompilationTest and QueryPlannerRuleOptionsTest.

Now JoinPushTransitivePredicatesRule is ready to ship, 6 tests in ResourceBasedQueryPlansTest are checked and updated to reflect addition of JoinPushTransitivePredicatesRule.

Now AggregateFunctionsPushdown is disabled by default, because enabling
it incurs large plan behavior changes.

AggregateJoinRemove is not effective rn because the round of prune
before BASIC_RULES would incur redundant Project and Exchange to rules.
The reason is still to be investigated.
This pruning would incurred ~53 changes in ResourceBasedQueryPlansTest.
So currently, the prune before BASIC_RULES is disabled, so is the test for AggregateJoinRemoveRule in QueryCompilationTest.